backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
openai
python-multipart
python-dotenv
pydantic
passlib
bcrypt
PyJWT
psycopg2-binary
